Updated the main.py
Added 2 alternate vectorization functions that can be manually switched between to compare results (TfidfVectorizer or CountVectorizer) as well as records and displays the time it took for the model to run. Also updated how the accuracy is displayed so now it goes in-depth over the number of true positives, false positives, true negatives, and false negatives as well as the overall accuracy.